Move SymmetrizationLevel into its own file
