+++ pyquil/experiment/_symmetrization.py
@@ -1,0 +1,24 @@
+##############################################################################
+# Copyright 2016-2019 Rigetti Computing
+#
+#    Licensed under the Apache License, Version 2.0 (the "License");
+#    you may not use this file except in compliance with the License.
+#    You may obtain a copy of the License at
+#
+#        http://www.apache.org/licenses/LICENSE-2.0
+#
+#    Unless required by applicable law or agreed to in writing, software
+#    distributed under the License is distributed on an "AS IS" BASIS,
+#    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#    See the License for the specific language governing permissions and
+#    limitations under the License.
+##############################################################################
+
+from enum import IntEnum
+
+class SymmetrizationLevel(IntEnum):
+    EXHAUSTIVE = -1
+    NONE = 0
+    OA_STRENGTH_1 = 1
+    OA_STRENGTH_2 = 2
+    OA_STRENGTH_3 = 3
